Hoang trung hieu

+++ server/src/main/java/com/springboot/bookstore/repository/OrderDetailsRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 
 
-public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Long> {
+public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Integer> {
     @Query("SELECT p.id, p.image, p.title, p.oldPrice, p.currentPrice, SUM(od.quantity) as totalQuantity " +
             "FROM OrderDetails od " +
             "JOIN od.product p " +
